[HIVE-FIX] Major HIVE-MIND collective update: Remove 31 JS files, fix 20+ compilation errors, add Windows gradle config, enhance security

+++ app/src/main/java/com/example/snacktrack/data/repository/AIRepository.kt
@@ -404,7 +404,7 @@ class AIRepository(
         // Calculate trend (kg per day)
         val trend = if (history.size >= 2) {
             (history.last().weight - history.first().weight) / 
-            java.time.temporal.ChronoUnit.DAYS.between(history.first().timestamp.toLocalDate(), history.last().timestamp.toLocalDate()).toInt()
+            java.time.temporal.ChronoUnit.DAYS.between(history.first().timestamp.toLocalDate(), history.last().timestamp.toLocalDate()).toDouble()
         } else 0.0
         
         // Adjust trend based on calorie differential
